% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_stall.R
\name{in_stall}
\alias{in_stall}
\title{Install locally from your own machine}
\usage{
in_stall(pkg, version = NULL, force = TRUE)
}
\arguments{
\item{pkg}{(character) A package name. Required.}

\item{version}{(character) If NULL, we try to install from the latest
version prior to your current version}

\item{force}{(logical) force installation attempt. Default: \code{TRUE}}
}
\description{
Install locally from your own machine
}
\examples{
\dontrun{
in_stall(pkg = "abind")
in_stall(pkg = "abind", version = "3.2")
}
}

